requirements.txt
fredapi==0.5.1
hmmlearn==0.3.0
keras==2.15.0
matplotlib==3.8.2
numpy==1.26.3
pandas==2.2.2
pyarrow==15.0.0
python-dotenv==1.0.1
scikit-learn==1.3.2
seaborn==0.13.2
statsmodels==0.14.1
tensorflow==2.15.0
tensorflow_macos==2.15.0
yfinance==0.2.37
ipython==8.19.0
ipywidgets==8.1.1
ipykernel==6.28.0
jupyter==1.0.0
jupyterlab==4.0.10